feat: Challenge Programs
Add challenge programs, and the proficiency matrix
When active, Challenge B keeps track of your
proficiency with each possible pair of instruments.

This proficiency is used in tandem with a user-selectable
program to calculate the pair you should work on next.

A time-trial mode is added to force you to determine the
notes used intuitively. Guessing successfully within the
time limit earns a point, guessing incorrectly, or taking
too long to guess loses a point.

Proficiency level increases with 5 points,
proficiency level decreases with -10 points.

The programs are separated into levels to offer
progressively more challenging / advanced practice
routines that build on the skills taught in the
earlier levels.

Voices have been carefully selected, especially
for early levels, so that you do not encounter
unfair situations with complex/misleading
harmonics early on.

Signed-off-by: NTOP <notalkonlyprox@gmail.com>